+++ InventoryManagement.Core/Models/Product.cs
@@ -14,8 +14,8 @@ namespace InventoryManagement.Models
         public int ColorId { get; set; }
         public Color Color { get; set; }
         public bool Gender { get; set; }
-        public double Price { get; set; }
-        public double SellingPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal SellingPrice { get; set; }
         public string ProductYear { get; set; }
         public int Size { get; set; }
         public int Count { get; set; }
