+++ InventoryManagement.Core/Models/SalesDetails.cs
@@ -8,7 +8,7 @@ namespace InventoryManagement.Models
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
         public int UserId { get; set; }
         public User User { get; set; }
